Move CodableFeedStore to it's own file
