+++ EssentialDeveloper/EssentialDeveloperTests/Feed Cache/CodableFeedStoreTests.swift
@@ -1,82 +1,6 @@
 import XCTest
 import EssentialDeveloper
 
-class CodableFeedStore: FeedStore {
-    private struct Cache: Codable {
-        let feed: [CodableFeedImage]
-        let timestamp: Date
-        
-        var localFeed: [LocalFeedImage] {
-            return feed.map { $0.local}
-        }
-    }
-    
-    private struct CodableFeedImage: Equatable, Codable {
-        private let id: UUID
-        private let description: String?
-        private let location: String?
-        private let url: URL
-        
-        init(
-            localFeedImage: LocalFeedImage
-        ) {
-            self.id = localFeedImage.id
-            self.description = localFeedImage.description
-            self.location = localFeedImage.location
-            self.url = localFeedImage.url
-        }
-        
-        var local: LocalFeedImage {
-            return LocalFeedImage(id: id, description: description, location: location, url: url)
-        }
-    }
-    
-    private let storeURL: URL
-    
-    init(storeURL: URL) {
-        self.storeURL = storeURL
-    }
-    
-    func retrieve(completion: @escaping RetrievalCompletion) {
-        guard let data = try? Data(contentsOf: storeURL) else {
-            return completion(.empty)
-        }
-        
-        do {
-            let decoder = JSONDecoder()
-            let cache = try decoder.decode(Cache.self, from: data)
-            completion(.found(feed: cache.localFeed, timestamp: cache.timestamp))
-        } catch {
-            completion(.failure(error))
-        }
-    }
-    
-    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        do {
-            let encoder = JSONEncoder()
-            let encoded = try encoder.encode(Cache(feed: feed.map(CodableFeedImage.init), timestamp: timestamp))
-            try encoded.write(to: storeURL)
-            completion(nil)
-        }
-        catch {
-            completion(error)
-        }
-    }
-    
-    func deleteCachedFeed(completion: @escaping DeletionCompletion){
-        guard FileManager.default.fileExists(atPath: storeURL.path) else {
-            return completion(nil)
-        }
-        
-        do {
-            try FileManager.default.removeItem(at: storeURL)
-            completion(nil)
-        } catch {
-            completion(error)
-        }
-    }
-}
-
 final class CodableFeedStoreTests: XCTestCase {
     override func setUp() {
         super.setUp()
